ПростаЗадачка4.cpp

+++ ПростаЗадачка4.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    double A, B, C;
+    float A, B, C;
 
     cout << "Введіть розмір A: ";
     cin >> A;
@@ -11,8 +11,8 @@ int main() {
     cout << "Введіть розмір C: ";
     cin >> C;
 
-    double surfaceArea = 2 * (A * B + A * C + B * C);
-    double volume = A * B * C;
+    float surfaceArea = 2 * (A * B + A * C + B * C);
+    float volume = A * B * C;
 
     cout << "S: " << surfaceArea << endl;
     cout << "V: " << volume << endl;
